Save pre-processor results.
Keep a local reference to the summary XmlDocument so we can push it back into the DataBlock.DocumentData object. Previously the reference was lost when preprocessing returned.

Move the pre-processing step to only run on Exports. (Pre-processing a Remove makes no sense, and would fail since there's no DocumentData to process.)